+++ eslint.config.ts
@@ -44,15 +44,9 @@ export default eslintConfig({
     casePolice: true,
     import: {
       ignores: TEST_DIR_GLOB,
-      overrides: {
-        'import/no-extraneous-dependencies': (_, options) => [
-          2,
-          {
-            ...options?.[0],
-            whitelist: ['import-meta-resolve'], // Patched + bundled
-          },
-        ],
-      },
+      extraneousDependenciesWhitelist: [
+        'import-meta-resolve', // Bundled (patched)
+      ],
     },
     perfectionist: {
       configSortObjects: {